pyproject.toml
[tool.poetry]
name = "annolab"
version = "0.1.0"
description = ""
authors = ["Muhammad Umair <49072162+itskanny@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pytz = "^2021.3"  # https://github.com/stub42/pytz
python-slugify = "^5.0.2"  # https://github.com/un33k/python-slugify
Pillow = "^9.0.1"  # https://github.com/python-pillow/Pillow
argon2-cffi = "^21.3.0"  # https://github.com/hynek/argon2_cffi
whitenoise = "^6.0.0"  # https://github.com/evansd/whitenoise
redis = "^4.1.3"  # https://github.com/redis/redis-py
celery = "^5.2.3"  # pyup: < 6.0  # https://github.com/celery/celery
django-celery-beat = "^2.2.1"  # https://github.com/celery/django-celery-beat
uvicorn = "^0.17.4"  # https://github.com/encode/uvicorn

# Django
# ------------------------------------------------------------------------------
django = "^3.2.12"  # pyup: < 4.0  # https://www.djangoproject.com/
django-environ = "^0.8.1"  # https://github.com/joke2k/django-environ
django-model-utils = "^4.2.0"  # https://github.com/jazzband/django-model-utils
django-allauth = "^0.48.0"  # https://github.com/pennersr/django-allauth
django-crispy-forms = "^1.14.0"  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5 = "^0.6"  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-redis = "^5.2.0"  # https://github.com/jazzband/django-redis
# Django REST Framework
djangorestframework = "^3.13.1"  # https://github.com/encode/django-rest-framework
django-cors-headers = "^3.11.0" # https://github.com/adamchainz/django-cors-headers
# DRF-spectacular for api documentation
drf-spectacular = "^0.21.2"
drf-yasg = "^1.20.0"
djoser = "^2.1.0"
django-useful = "^0.1.2"
drf-extensions = "^0.7.1"


[tool.poetry.dev-dependencies]
Werkzeug = "^2.0.3" # https://github.com/pallets/werkzeug
ipdb = "^0.13.9"  # https://github.com/gotcha/ipdb
psycopg2-binary = "^2.9.3"  # https://github.com/psycopg/psycopg2
watchgod = "^0.7"  # https://github.com/samuelcolvin/watchgod

# Testing
# ------------------------------------------------------------------------------
mypy = "^0.931"  # https://github.com/python/mypy
django-stubs = "^1.9.0"  # https://github.com/typeddjango/django-stubs
pytest = "^7.0.1"  # https://github.com/pytest-dev/pytest
pytest-sugar = "^0.9.4"  # https://github.com/Frozenball/pytest-sugar
djangorestframework-stubs = "^1.4.0"  # https://github.com/typeddjango/djangorestframework-stubs

# Documentation
# ------------------------------------------------------------------------------
sphinx = "^4.4.0"  # https://github.com/sphinx-doc/sphinx
sphinx-autobuild = "^2021.3.14" # https://github.com/GaretJax/sphinx-autobuild

# Code quality
# ------------------------------------------------------------------------------
flake8 = "^4.0.1"  # https://github.com/PyCQA/flake8
flake8-isort = "^4.1.1"  # https://github.com/gforcada/flake8-isort
coverage = "^6.3.1"  # https://github.com/nedbat/coveragepy
black = "^22.1.0"  # https://github.com/psf/black
pylint-django = "^2.5.0"  # https://github.com/PyCQA/pylint-django
pylint-celery = "^0.3"  # https://github.com/PyCQA/pylint-celery
pre-commit = "^2.17.0"  # https://github.com/pre-commit/pre-commit

# Django
# ------------------------------------------------------------------------------
factory-boy = "^3.2.1"  # https://github.com/FactoryBoy/factory_boy

django-debug-toolbar = "^3.2.4"  # https://github.com/jazzband/django-debug-toolbar
django-extensions = "^3.1.5"  # https://github.com/django-extensions/django-extensions
django-coverage-plugin = "^2.0.2"  # https://github.com/nedbat/django_coverage_plugin
pytest-django = "^4.5.2"  # https://github.com/pytest-dev/pytest-django


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
